requirements.txt
streamlit>=1.26.0
spacy>=3.6.0
nltk>=3.8.1
PyPDF2>=3.0.0
python-docx>=0.8.11
reportlab>=3.6.12
requests>=2.28.2
openai>=1.3.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pdfkit>=1.0.0
matplotlib>=3.7.2
wordcloud>=1.8.2.2
pandas>=2.0.3
